Removed "Classes" folder. Created individual files for each element type. Extracted Element and ElementSeed classes into their element type's file.
